+++ src/debugger/debugger_fscript.c
@@ -19,15 +19,70 @@
  *
  */
 
+#include "tkc/object_default.h"
 #include "debugger/debugger_server.h"
 #include "debugger/debugger_message.h"
 #include "debugger/debugger_fscript.h"
+
+typedef struct _call_stack_frame_t {
+  const char* name;
+  tk_object_t* locals;
+} call_stack_frame_t;
+
+static call_stack_frame_t* call_stack_frame_create(const char* name, tk_object_t* locals) {
+  call_stack_frame_t* frame = TKMEM_ZALLOC(call_stack_frame_t);
+  return_value_if_fail(frame != NULL, NULL);
+
+  frame->name = name;
+  frame->locals = locals;
+
+  return frame;
+}
+
+static ret_t call_stack_frame_destroy(call_stack_frame_t* frame) {
+  return_value_if_fail(frame != NULL, RET_BAD_PARAMS);
+
+  TKMEM_FREE(frame);
+
+  return RET_OK;
+}
+
+static ret_t debugger_fscript_enter_func(debugger_t* debugger, const char* name,
+                                         tk_object_t* locals) {
+  call_stack_frame_t* frame = NULL;
+  debugger_fscript_t* d = DEBUGGER_FSCRIPT(debugger);
+  return_value_if_fail(d != NULL, RET_BAD_PARAMS);
+  return_value_if_fail(name != NULL, RET_BAD_PARAMS);
+
+  frame = call_stack_frame_create(name, locals);
+  return_value_if_fail(frame != NULL, RET_OOM);
+
+  if (darray_push(&(d->call_stack_frames), frame) != RET_OK) {
+    call_stack_frame_destroy(frame);
+    return RET_OOM;
+  } else {
+    log_debug("enter %s\n", name);
+    return RET_OK;
+  }
+}
+
+static ret_t debugger_fscript_leave_func(debugger_t* debugger) {
+  call_stack_frame_t* frame = NULL;
+  debugger_fscript_t* d = DEBUGGER_FSCRIPT(debugger);
+  return_value_if_fail(d != NULL, RET_BAD_PARAMS);
+
+  frame = (call_stack_frame_t*)darray_pop(&(d->call_stack_frames));
+  return_value_if_fail(frame != NULL, RET_BAD_PARAMS);
+
+  log_debug("leave %s\n", frame->name);
+
+  return call_stack_frame_destroy(frame);
+}
 
 static ret_t debugger_fscript_lock(debugger_t* debugger) {
   debugger_fscript_t* d = DEBUGGER_FSCRIPT(debugger);
   return_value_if_fail(d != NULL, RET_BAD_PARAMS);
 
-  log_debug("debugger_fscript_lock\n");
   return tk_mutex_nest_lock(d->mutex);
 }
 
@@ -35,7 +90,6 @@ static ret_t debugger_fscript_unlock(debugger_t* debugger) {
   debugger_fscript_t* d = DEBUGGER_FSCRIPT(debugger);
   return_value_if_fail(d != NULL, RET_BAD_PARAMS);
 
-  log_debug("debugger_fscript_unlock\n");
   return tk_mutex_nest_unlock(d->mutex);
 }
 
@@ -92,6 +146,14 @@ static bool_t debugger_fscript_is_paused(debugger_t* debugger) {
   return ret;
 }
 
+static ret_t debugger_fscript_clear_step_stops(debugger_fscript_t* d) {
+  d->next_stop_line = -1;
+  d->next_stop_executed_line = -1;
+  d->next_stop_call_frame_index = -1;
+
+  return RET_OK;
+}
+
 static ret_t debugger_fscript_next(debugger_t* debugger) {
   ret_t ret = RET_FAIL;
   debugger_fscript_t* d = DEBUGGER_FSCRIPT(debugger);
@@ -100,7 +162,9 @@ static ret_t debugger_fscript_next(debugger_t* debugger) {
   if (debugger_fscript_lock(debugger) == RET_OK) {
     if (d->fscript != NULL && d->paused) {
       ret = RET_OK;
+      debugger_fscript_clear_step_stops(d);
       d->next_stop_executed_line = d->executed_lines + 1;
+      log_debug("debugger_fscript_next: %d %d\n", d->executed_lines, d->next_stop_executed_line);
     }
     debugger_fscript_unlock(debugger);
 
@@ -113,18 +177,69 @@ static ret_t debugger_fscript_next(debugger_t* debugger) {
 }
 
 static ret_t debugger_fscript_step_in(debugger_t* debugger) {
-  /*TODO*/
-  return debugger_fscript_next(debugger);
+  ret_t ret = RET_FAIL;
+  debugger_fscript_t* d = DEBUGGER_FSCRIPT(debugger);
+  return_value_if_fail(d != NULL, RET_BAD_PARAMS);
+
+  if (debugger_fscript_lock(debugger) == RET_OK) {
+    if (d->fscript != NULL && d->paused) {
+      ret = RET_OK;
+      debugger_fscript_clear_step_stops(d);
+      d->next_stop_call_frame_index = d->call_stack_frames.size + 1;
+      log_debug("set step in:%d\n", d->next_stop_call_frame_index);
+    }
+    debugger_fscript_unlock(debugger);
+
+    if (ret == RET_OK) {
+      tk_cond_var_awake(d->cond_var);
+    }
+  }
+
+  return ret;
 }
 
 static ret_t debugger_fscript_step_out(debugger_t* debugger) {
-  /*TODO*/
-  return debugger_fscript_next(debugger);
+  ret_t ret = RET_FAIL;
+  debugger_fscript_t* d = DEBUGGER_FSCRIPT(debugger);
+  return_value_if_fail(d != NULL, RET_BAD_PARAMS);
+  return_value_if_fail(d->call_stack_frames.size > 0, RET_BAD_PARAMS);
+
+  if (debugger_fscript_lock(debugger) == RET_OK) {
+    if (d->fscript != NULL && d->paused) {
+      ret = RET_OK;
+      debugger_fscript_clear_step_stops(d);
+      d->next_stop_call_frame_index = d->call_stack_frames.size - 1;
+      log_debug("set step out:%d\n", d->next_stop_call_frame_index);
+    }
+    debugger_fscript_unlock(debugger);
+
+    if (ret == RET_OK) {
+      tk_cond_var_awake(d->cond_var);
+    }
+  }
+
+  return ret;
 }
 
 static ret_t debugger_fscript_step_over(debugger_t* debugger) {
-  /*TODO*/
-  return debugger_fscript_next(debugger);
+  ret_t ret = RET_FAIL;
+  debugger_fscript_t* d = DEBUGGER_FSCRIPT(debugger);
+  return_value_if_fail(d != NULL, RET_BAD_PARAMS);
+
+  if (debugger_fscript_lock(debugger) == RET_OK) {
+    if (d->fscript != NULL && d->paused) {
+      ret = RET_OK;
+      debugger_fscript_clear_step_stops(d);
+      d->next_stop_line = d->prev_breaked_line + 1;
+    }
+    debugger_fscript_unlock(debugger);
+
+    if (ret == RET_OK) {
+      tk_cond_var_awake(d->cond_var);
+    }
+  }
+
+  return ret;
 }
 
 static ret_t debugger_fscript_continue(debugger_t* debugger) {
@@ -135,8 +250,7 @@ static ret_t debugger_fscript_continue(debugger_t* debugger) {
   if (debugger_fscript_lock(debugger) == RET_OK) {
     if (d->fscript != NULL && d->paused) {
       ret = RET_OK;
-      d->next_stop_line = -1;
-      d->next_stop_executed_line = -1;
+      debugger_fscript_clear_step_stops(d);
     }
     debugger_fscript_unlock(debugger);
 
@@ -159,10 +273,17 @@ static ret_t debugger_fscript_get_code(debugger_t* debugger, binary_data_t* code
 }
 
 tk_object_t* debugger_fscript_get_local(debugger_t* debugger, uint32_t frame_index) {
+  uint32_t index = 0;
+  call_stack_frame_t* frame = NULL;
   debugger_fscript_t* d = DEBUGGER_FSCRIPT(debugger);
   return_value_if_fail(d != NULL && d->fscript != NULL, NULL);
-  /*FIXME: 目前只支持当前*/
-  return TK_OBJECT_REF(d->fscript->locals);
+  return_value_if_fail(d->call_stack_frames.size > frame_index, NULL);
+
+  index = d->call_stack_frames.size - frame_index - 1;
+  frame = (call_stack_frame_t*)darray_get(&(d->call_stack_frames), index);
+  return_value_if_fail(frame != NULL, NULL);
+
+  return TK_OBJECT_REF(frame->locals);
 }
 
 tk_object_t* debugger_fscript_get_self(debugger_t* debugger) {
@@ -177,8 +298,15 @@ tk_object_t* debugger_fscript_get_global(debugger_t* debugger) {
 }
 
 static ret_t debugger_fscript_get_callstack(debugger_t* debugger, binary_data_t* callstack) {
+  uint32_t i = 0;
   debugger_fscript_t* d = DEBUGGER_FSCRIPT(debugger);
   return_value_if_fail(d != NULL && d->fscript != NULL, RET_BAD_PARAMS);
+
+  str_clear(&(d->callstack));
+  for (i = 0; i < d->call_stack_frames.size; i++) {
+    call_stack_frame_t* iter = (call_stack_frame_t*)darray_get(&(d->call_stack_frames), i);
+    str_append_more(&(d->callstack), iter->name, "\n", NULL);
+  }
 
   callstack->data = d->callstack.str;
   callstack->size = d->callstack.size + 1;
@@ -193,8 +321,7 @@ static ret_t debugger_fscript_clear_break_points(debugger_t* debugger) {
   return_value_if_fail(d != NULL, RET_BAD_PARAMS);
 
   if (debugger_fscript_lock(debugger) == RET_OK) {
-    d->next_stop_line = -1;
-    d->next_stop_executed_line = -1;
+    debugger_fscript_clear_step_stops(d);
     ret = darray_clear(&(d->break_points));
 
     paused = d->paused;
@@ -261,8 +388,8 @@ static ret_t debugger_fscript_update_code(debugger_t* debugger, const binary_dat
 
   if (debugger_fscript_lock(debugger) == RET_OK) {
     ret = RET_OK;
+    d->code_changed = TRUE;
     str_set_with_len(&(d->code), (char*)(code->data), code->size);
-    /*TODO*/
     debugger_fscript_unlock(debugger);
   }
 
@@ -314,6 +441,7 @@ static ret_t debugger_fscript_on_destroy(tk_object_t* obj) {
   str_reset(&(debugger->code));
   str_reset(&(debugger->callstack));
   darray_deinit(&(debugger->break_points));
+  darray_deinit(&(debugger->call_stack_frames));
 
   return RET_OK;
 }
@@ -337,6 +465,7 @@ debugger_t* debugger_fscript_create(void) {
   str_init(&(debugger->code), 100);
   str_init(&(debugger->callstack), 100);
   darray_init(&(debugger->break_points), 10, NULL, NULL);
+  darray_init(&(debugger->call_stack_frames), 10, (tk_destroy_t)call_stack_frame_destroy, NULL);
 
   return (debugger_t*)debugger;
 }
@@ -382,32 +511,80 @@ ret_t debugger_fscript_set_fscript(debugger_t* debugger, fscript_t* fscript) {
   debugger_fscript_t* d = DEBUGGER_FSCRIPT(debugger);
   return_value_if_fail(d != NULL, RET_BAD_PARAMS);
 
-  d->fscript = fscript;
   if (fscript) {
+
+    if (d->code_changed) {
+      d->code_changed = FALSE;
+      return_value_if_fail(fscript_reload(fscript, d->code.str) == RET_OK, RET_FAIL);
+    }
+
+    d->fscript = fscript;
     d->executed_lines = 0;
     d->last_executed_line = 0;
-    d->next_stop_line = -1;
     d->prev_breaked_line = -1;
-    d->next_stop_executed_line = -1;
-    str_set(&(d->callstack), "<root>\n");
+    debugger_fscript_clear_step_stops(d);
+
     fscript_set_print_func(fscript, debugger_fscript_print_func);
     fscript_set_on_error(fscript, debugger_fscript_on_error, d);
+
+    if (fscript->locals == NULL) {
+      /*fscript本身是延迟创建locals的，这会导致debugger拿不到locals对象，所以提前创建它。*/
+      fscript->locals = object_default_create();
+    }
+
+    darray_clear(&(d->call_stack_frames));
+    debugger_fscript_enter_func(debugger, "<root>", fscript->locals);
   } else {
+    debugger_fscript_leave_func(debugger);
+    fscript_set_print_func(d->fscript, NULL);
+    fscript_set_on_error(d->fscript, NULL, NULL);
     emitter_dispatch_simple_event(EMITTER(debugger), DEBUGGER_RESP_MSG_COMPLETED);
+    d->fscript = NULL;
   }
 
   return RET_OK;
 }
 
+static bool_t debugger_fscript_should_pause(debugger_fscript_t* d, int32_t line) {
+  bool_t paused = darray_find_index(&(d->break_points), tk_pointer_from_int(line)) >= 0;
+
+  if (paused) {
+    log_debug("break at line:%d\n", line);
+  }
+
+  if (d->next_stop_executed_line >= 0 && d->executed_lines >= d->next_stop_executed_line) {
+    log_debug("break for step next:next_stop_executed_line=%d line=%d\n",
+              d->next_stop_executed_line, line);
+    paused = TRUE;
+  }
+
+  if (line == d->next_stop_line) {
+    log_debug("break for over:next_stop_line=%d line=%d\n", d->next_stop_line, line);
+    paused = TRUE;
+  }
+
+  if (d->next_stop_call_frame_index == d->call_stack_frames.size) {
+    log_debug("break for step in:%d\n", d->call_stack_frames.size);
+    paused = TRUE;
+  }
+
+  d->paused = paused;
+
+  if (d->prev_breaked_line == line) {
+    paused = FALSE;
+    log_debug("d->prev_breaked_line == line\n");
+  } else {
+    d->prev_breaked_line = -1;
+  }
+
+  return paused;
+}
+
 static ret_t debugger_fscript_before_exec_func(debugger_t* debugger, int32_t line) {
-  bool_t paused = FALSE;
   debugger_fscript_t* d = DEBUGGER_FSCRIPT(debugger);
 
   if (debugger_fscript_lock(debugger) == RET_OK) {
-    bool_t bp = darray_find_index(&(d->break_points), tk_pointer_from_int(line)) >= 0;
-
-    d->paused = bp || line == d->next_stop_line || d->executed_lines == d->next_stop_executed_line;
-    paused = d->paused && (d->prev_breaked_line != line);
+    bool_t paused = debugger_fscript_should_pause(d, line);
 
     if (paused) {
       debugger_breaked_event_t event;
@@ -418,8 +595,9 @@ static ret_t debugger_fscript_before_exec_func(debugger_t* debugger, int32_t lin
     debugger_fscript_unlock(debugger);
 
     if (paused) {
-      log_debug("debugger wait for instruction\n");
+      log_debug("debugger wait for instruction begin\n");
       tk_cond_var_wait(d->cond_var, 0xffffff);
+      log_debug("debugger wait for instruction done\n");
     }
   }
 
@@ -455,9 +633,38 @@ ret_t debugger_fscript_exec_func(fscript_t* fscript, const char* name, fscript_f
   } else {
     int32_t line = iter->row;
 
+    /*name不为NUL则是脚本函数*/
+    if (name != NULL) {
+      if (debugger_fscript_lock(debugger) == RET_OK) {
+        debugger_fscript_enter_func(debugger, name, fscript->locals);
+        debugger_fscript_unlock(debugger);
+      }
+    }
+
     debugger_fscript_before_exec_func(debugger, line);
     ret = fscript_exec_func_default(fscript, iter, result);
     debugger_fscript_after_exec_func(debugger, line);
+
+    if (name != NULL) {
+      if (debugger_fscript_lock(debugger) == RET_OK) {
+        bool_t paused = FALSE;
+        debugger_fscript_t* d = DEBUGGER_FSCRIPT(debugger);
+        debugger_fscript_leave_func(debugger);
+
+        if (d->next_stop_call_frame_index == d->call_stack_frames.size) {
+          paused = TRUE;
+          d->paused = paused;
+          log_debug("stop for step out\n");
+        }
+        debugger_fscript_unlock(debugger);
+
+        if (paused) {
+          log_debug("debugger wait for instruction begin\n");
+          tk_cond_var_wait(d->cond_var, 0xffffff);
+          log_debug("debugger wait for instruction done\n");
+        }
+      }
+    }
   }
 
   return ret;